refactor: Location VO를 global 패키지로 이동 #8
