--- src/main/java/com/example/mohago_nocar/festival/domain/model/vo/Location.java
@@ -1,24 +1,0 @@
-package com.example.mohago_nocar.festival.domain.model.vo;
-
-import jakarta.persistence.Embeddable;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class Location {
-
-    private Double latitude;
-    private Double longitude;
-
-    public static Location from(Double latitude, Double longitude) {
-        return Location.builder()
-                .latitude(latitude)
-                .longitude(longitude)
-                .build();
-    }
-}
